[nemesys/modelling/centres/pytorch_centre.py] Refactor for multiple inheritance
Because a PyTorchCentre is a multiple inheritance object (uses torch.nn.Module
and nemesys.modelling.centres.Centre), one super init call won't cut it.
Furthermore, it is necessary to first initialize the torch.nn.Module so it
can record the Centre components after.

Therefore, we explicitly call the super init methods in the wanted order.